pyproject.toml
[project]
name = "bootcamp"
version = "0.1.0"
description = "Materials for the MLExpert.io Bootcamp"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastparquet>=2024.11.0",
    "jupyterlab>=4.3.4",
    "litellm>=1.59.8",
    "loguru>=0.7.3",
    "matplotlib>=3.10.0",
    "numpy>=2.2.2",
    "ollama>=0.4.7",
    "openai>=1.60.1",
    "opik>=1.4.5",
    "pandas>=2.2.3",
    "pre-commit>=4.1.0",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.3",
    "seaborn>=0.13.2",
    "tabulate>=0.9.0",
    "tiktoken>=0.8.0",
    "torch>=2.5.1",
    "transformers>=4.48.1",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[tool.setuptools]
packages = ["bootcamp"]
